connectors/bright_data_dataset/requirements.txt
brightdata_sdk==1.1.3
python-dotenv==1.2.1
pyyaml==6.0.3